build(extension): migrate from Vite/CRXJS to WXT framework
Why:
- The previous setup using `@crxjs/vite-plugin` was incompatible with Vite 6+ (Rolldown engine), resulting in build crashes (`Missing field moduleType`).
- A transition to WXT was required to ensure a stable, future-proof, and DX-friendly build pipeline tailored for Chrome Extensions (Manifest V3).

What:
- Removed `vite` and `@crxjs/vite-plugin` dependencies from the extension package.
- Installed `wxt` and `@wxt-dev/module-react`.
- Restructured the source files into the WXT `entrypoints/` directory convention (e.g., `background.ts` and `sidepanel/`).
- Replaced `vite.config.ts` and `manifest.json` with a unified `wxt.config.ts` configuration.
- Updated `tsconfig.json` to extend WXT's auto-generated types and removed legacy TS configs.
- Updated `package.json` scripts to use WXT CLI commands.
- Added the new development extension ID (`*.chromiumapp.org`) to `supabase/config.toml` for authentication redirects.
