+++ apps/extension/entrypoints/sidepanel/main.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import SidePanel from "./SidePanel";
-import "./index.css";
+import SidePanel from "../../src/SidePanel";
+import "../../src/index.css";
 
 // biome-ignore lint/style/noNonNullAssertion: root element exists in index.html
 ReactDOM.createRoot(document.getElementById("root")!).render(
