--- apps/extension/manifest.json
@@ -1,40 +1,0 @@
-{
-	"manifest_version": 3,
-	"name": "sitecue",
-	"version": "1.1.0",
-	"description": "Leave notes linked to websites in the side panel. A simple way to keep the notes you need exactly where you need them.",
-	"action": {
-		"default_title": "Open sitecue"
-	},
-	"icons": {
-		"16": "sitecue_icon16.png",
-		"32": "sitecue_icon32.png",
-		"48": "sitecue_icon48.png",
-		"128": "sitecue_icon128.png"
-	},
-	"background": {
-		"service_worker": "src/background.ts",
-		"type": "module"
-	},
-	"side_panel": {
-		"default_path": "sidepanel.html"
-	},
-	"permissions": [
-		"sidePanel",
-		"activeTab",
-		"storage",
-		"tabs",
-		"identity",
-		"scripting"
-	],
-	"optional_host_permissions": ["<all_urls>"],
-	"host_permissions": ["http://localhost:8787/*", "http://127.0.0.1:54321/*"],
-	"commands": {
-		"_execute_action": {
-			"suggested_key": {
-				"default": "Ctrl+Shift+E",
-				"mac": "Command+Shift+E"
-			}
-		}
-	}
-}
