+++ apps/extension/package.json
@@ -4,11 +4,11 @@
 	"version": "0.0.0",
 	"type": "module",
 	"scripts": {
-		"dev": "vite",
-		"build": "tsc -b && vite build",
+		"dev": "wxt",
+		"build": "wxt build",
+		"build:zip": "wxt zip",
 		"lint": "biome check .",
 		"lint:fix": "biome check --write .",
-		"preview": "vite preview",
 		"test": "vitest",
 		"test:ui": "vitest --ui"
 	},
@@ -27,7 +27,6 @@
 		"tailwind-merge": "^3.4.0"
 	},
 	"devDependencies": {
-		"@crxjs/vite-plugin": "^2.4.0",
 		"@tailwindcss/vite": "^4.1.18",
 		"@testing-library/jest-dom": "^6.9.1",
 		"@testing-library/react": "^16.3.2",
@@ -37,13 +36,13 @@
 		"@types/react": "^19.2.5",
 		"@types/react-dom": "^19.2.3",
 		"@types/react-syntax-highlighter": "^15.5.13",
-		"@vitejs/plugin-react": "^6.0.1",
+		"@wxt-dev/module-react": "^1.2.2",
 		"autoprefixer": "^10.4.23",
 		"happy-dom": "^20.8.4",
 		"postcss": "^8.5.6",
 		"tailwindcss": "^4.1.18",
 		"typescript": "~5.9.3",
-		"vite": "7.2.4",
-		"vitest": "^4.1.0"
+		"vitest": "^4.1.0",
+		"wxt": "^0.20.20"
 	}
 }
