+++ apps/extension/tsconfig.json
@@ -1,7 +1,3 @@
 {
-	"files": [],
-	"references": [
-		{ "path": "./tsconfig.app.json" },
-		{ "path": "./tsconfig.node.json" }
-	]
+  "extends": "./.wxt/tsconfig.json"
 }
